+++ digestor/sdss.py
@@ -9,11 +9,14 @@ Convert SDSS SQL (MS SQL Server) table definitions to Data Lab SQL (PostgreSQL).
 import os
 import re
 import sys
+import time
 # from datetime import datetime
 from argparse import ArgumentParser
 
 from pkg_resources import resource_filename
 # from pytz import utc
+import numpy as np
+from astropy.table import Table
 
 from .base import Digestor
 
@@ -26,7 +29,6 @@ class SDSS(Digestor):
     #
     _SQLre = {'comment': re.compile(r'\s*--/(H|T)\s+(.*)$'),
               'column': re.compile(r'\s*(\S+)\s+(\S+)\s*([^,]+),\s*(.*)$')}
-
     #
     # Map SQL Server data types to PostgreSQL.
     #
@@ -37,9 +39,14 @@ class SDSS(Digestor):
     # Ignore columns that are specific to the SDSS CAS system.
     #
     _skip_columns = ('htmid', 'loadversion')
+    #
+    # Identify columns that contain photometric flags
+    #
+    _flagre = re.compile(r'flags(|_[ugriz])$', re.I)
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
+        self.NOFITS = dict()
 
     def parseSQL(self, filename):
         """Parse an entire SQL file.
@@ -145,7 +152,7 @@ class SDSS(Digestor):
                 r = data[i + 5:j].strip()
                 if m == 'F':
                     if ' ' in r:
-                        r = "{0}[{1}]".format(*(r.split(' ')))
+                        r = "{0}[{1}]".format(*(re.split(r'\s+', r)))
                     r = r.upper()
                     if r == 'NOFITS':
                         log.warning("Column %s is not defined in the corresponding FITS file!", column)
@@ -161,6 +168,46 @@ class SDSS(Digestor):
                     rename = r
         return (p, rename)
 
+    def fixNOFITS(self, filename):
+        """Fix any missing data designated by ``--/F NOFITS`` using
+        the YAML configuration file `filename`.
+
+        Parameters
+        ----------
+        filename : :class:`str`
+            Name of the YAML configuration file.
+        """
+        log = self.logName('sdss.SDSS.fixNOFITS')
+        config = self._getYAML(filename)
+        if config is not None:
+            try:
+                self.NOFITS = config[self.schema][self.table]['NOFITS']
+            except KeyError:
+                log.debug("No instructions found.")
+        return
+
+    def fixMapping(self, filename):
+        """Fix any FITS to SQL mapping problems using the YAML configuration
+        file `filename`.
+
+        Parameters
+        ----------
+        filename : :class:`str`
+            Name of the YAML configuration file.
+        """
+        log = self.logName('sdss.SDSS.fixMapping')
+        config = self._getYAML(filename)
+        if config is not None:
+            try:
+                mapping = config[self.schema][self.table]['mapping']
+            except KeyError:
+                log.debug("No mappings found.")
+                return
+            for sc in mapping:
+                log.debug("self.mapping['%s'] = '%s'", sc, mapping[sc])
+                self.mapping[sc] = mapping[sc]
+        return
+
     def mapColumns(self):
         """Complete mapping of FITS table columns to SQL columns.
 
@@ -170,6 +217,7 @@ class SDSS(Digestor):
             If an expected mapping cannot be found.
         """
         log = self.logName('sdss.SDSS.mapColumns')
+        drop = list()
         for sc in self.colNames:
             if sc in self.mapping:
                 #
@@ -193,7 +241,7 @@ class SDSS(Digestor):
                     for fc in self.FITS:
                         for fcl in (fc.lower(), fc.lower().replace('_', ''),):
                             if fcl == mc.lower():
-                                log.debug("FITS: %s -> SQL: %s", fc, sc)
+                                log.debug("FITS: %s%s -> SQL: %s", fc, index, sc)
                                 self.mapping[sc] = fc + index
                                 verify_mapping = True
                 if not verify_mapping:
@@ -213,10 +261,27 @@ class SDSS(Digestor):
                 if sc == 'random_id':
                     log.info("Skipping %s which will be added by FITS2DB.",
                              sc)
+                elif sc in self.NOFITS:
+                    if self.NOFITS[sc] == 'drop':
+                        log.info("Dropping %s as requested.", sc)
+                        drop.append(sc)
+                    elif self.NOFITS[sc] == 'defer':
+                        log.info("Column %s will be added in post-processing.", sc)
+                    else:
+                        msg = "Unknown NOFITS instruction: %s!"
+                        log.error(msg, sc)
+                        raise KeyError(msg % sc)
                 else:
                     msg = "Could not find a FITS column corresponding to %s!"
                     log.error(msg, sc)
                     raise KeyError(msg % sc)
+        #
+        # Remove SQL columns that were requested to be dropped.
+        #
+        for sc in drop:
+            i = self.columnIndex(sc)
+            log.debug("del self.tapSchema['columns'][%d]", i)
+            del self.tapSchema['columns'][i]
         #
         # Check for FITS columns that are NOT mapped to the SQL file.
         #
@@ -230,6 +295,212 @@ class SDSS(Digestor):
                 else:
                     log.warning("FITS column %s will be dropped from SQL!", col)
         return
+
+    def _photoFlag(self, column, table):
+        """Handle photometric flags in SDSS data.
+
+        Parameters
+        ----------
+        column : :class:`dict`
+            A TapSchema column definition.
+        table : :class:`astropy.table.Table`
+            Table containing the input data.
+
+        Returns
+        -------
+        :class:`numpy.ndarray`
+            The combined flags and flags2 data, or ``None`` if the
+            column did not match.
+
+        Raises
+        ------
+        :exc:`AssertionError`
+            If the required columns are not present in the FITS file.
+        """
+        log = self.logName('sdss.SDSS._photoFlag')
+        m = self._flagre.match(column['column_name'])
+        if m is not None:
+            g = m.groups()[0].replace('_', '')
+            if g:
+                #
+                # Ensure FLAGS and FLAGS2 are present.
+                #
+                band = 'ugriz'.index(g)
+                assert column['datatype'] == 'bigint'
+                assert self.mapping[column['column_name']].lower() == 'flags[{0:d}]'.format(band)
+                assert 'FLAGS' in table.colnames
+                assert 'FLAGS2' in table.colnames
+                log.debug("np.left_shift(table['FLAGS2'][:, %d].astype(np.int64), 32) | table['FLAGS'][:, %d].astype(np.int64)", band, band)
+                return (np.left_shift(table['FLAGS2'][:, band].astype(np.int64), 32) |
+                        table['FLAGS'][:, band].astype(np.int64))
+            else:
+                #
+                # Ensure OBJC_FLAGS and OBJC_FLAGS2 are present.
+                #
+                assert column['datatype'] == 'bigint'
+                assert self.mapping[column['column_name']].lower() == 'objc_flags'
+                assert 'OBJC_FLAGS' in table.colnames
+                assert 'OBJC_FLAGS2' in table.colnames
+                log.debug("np.left_shift(table['OBJC_FLAGS2'].astype(np.int64), 32) | table['OBJC_FLAGS'].astype(np.int64)")
+                return (np.left_shift(table['OBJC_FLAGS2'].astype(np.int64), 32) |
+                        table['OBJC_FLAGS'].astype(np.int64))
+        return None
+
+    def processFITS(self, hdu=1, overwrite=False):
+        """Convert a pre-processed FITS file into one ready for database loading.
+
+        Parameters
+        ----------
+        hdu : :class:`int`, optional
+            Read data from this HDU (default 1).
+        overwrite : :class:`bool`, optional
+            If ``True``, remove any existing file.
+
+        Returns
+        -------
+        :class:`str`
+            The name of the file written.
+
+        Raises
+        ------
+        :exc:`ValueError`
+            If the FITS data type cannot be converted to SQL.
+        """
+        log = self.logName('sdss.SDSS.processFITS')
+        out = "{0.schema}.{0.table}.fits".format(self)
+        if os.path.exists(out) and not overwrite:
+            log.info("Using existing file: %s.", out)
+            return out
+        if os.path.exists(out):
+            log.info("Removing existing file: %s.", out)
+            os.remove(out)
+        type_map = {'bigint': ('K', 'J', 'I', 'B'),
+                    'integer': ('J', 'I', 'B'),
+                    'smallint': ('I', 'B'),
+                    'double': ('D', 'E'),
+                    'real': ('E',),
+                    'character': ('A',)}
+        np_map = {'bigint': np.int64,
+                  'integer': np.int32,
+                  'smallint': np.int16,
+                  'double': np.float64,
+                  'real': np.float32}
+        safe_conversion = {('J', 'smallint'): 2**15,
+                           ('A', 'smallint'): 2**15,
+                           ('A', 'integer'): 2**31,
+                           ('A', 'bigint'): 2**63}
+        rebase = re.compile(r'^(\d+)(\D+)')
+        columns = [c for c in self.tapSchema['columns']
+                   if c['table_name'] == self.table]
+        old = Table.read(self._inputFITS, hdu=hdu)
+        new = Table()
+        for col in columns:
+            if col['column_name'] == 'random_id':
+                log.info("Creating %s column using numpy.random.random().",
+                         col['column_name'])
+                stime = int(time.time())
+                log.debug('np.random.seed(%s)', stime)
+                np.random.seed(stime)
+                log.debug("new['%s'] = np.random.random((%d,)).astype(%s)",
+                          col['column_name'], len(old), str(np_map[col['datatype']]))
+                new[col['column_name']] = 100.0*np.random.random((len(old),)).astype(np_map[col['datatype']])
+                continue
+            if col['column_name'] in self.NOFITS:
+                log.info("Creating placeholder column %s for post-processing.",
+                         col['column_name'])
+                log.debug("new['%s'] = np.zeros((%d,), dtype=%s)",
+                          col['column_name'], len(old), str(np_map[col['datatype']]))
+                new[col['column_name']] = np.zeros((len(old),), dtype=np_map[col['datatype']])
+                continue
+            if 'flags' in col['column_name']:
+                flags64 = self._photoFlag(col, old)
+                if flags64 is not None:
+                    log.info("Combining photo flags for %s", col['column_name'])
+                    new[col['column_name']] = flags64
+                    continue
+            fcol = self.mapping[col['column_name']]
+            index = None
+            if '[' in fcol:
+                foo = fcol.split('[')
+                fcol = foo[0]
+                index = int(foo[1].strip(']'))
+            ftype = self.FITS[fcol]
+            fbasetype = rebase.sub(r'\2', ftype)
+            if fbasetype == type_map[col['datatype']][0]:
+                log.debug("Type match for %s -> %s.", fcol, col['column_name'])
+                if index is not None:
+                    log.debug("new['%s'] = old['%s'][:, %d]",
+                              col['column_name'], fcol, index)
+                    new[col['column_name']] = old[fcol][:, index]
+                else:
+                    log.debug("new['%s'] = old['%s']", col['column_name'], fcol)
+                    new[col['column_name']] = old[fcol]
+            elif fbasetype in type_map[col['datatype']]:
+                log.debug("Safe type conversion possible for %s (%s) -> %s (%s).",
+                          fcol, fbasetype, col['column_name'], col['datatype'])
+                if index is not None:
+                    log.debug("new['%s'] = old['%s'][:, %d].astype(%s)",
+                              col['column_name'], fcol, index,
+                              str(np_map[col['datatype']]))
+                    new[col['column_name']] = old[fcol][:, index].astype(np_map[col['datatype']])
+                else:
+                    log.debug("new['%s'] = old['%s'].astype(%s)",
+                              col['column_name'], fcol,
+                              str(np_map[col['datatype']]))
+                    new[col['column_name']] = old[fcol].astype(np_map[col['datatype']])
+            else:
+                if (fbasetype, col['datatype']) in safe_conversion:
+                    limit = safe_conversion[(fbasetype, col['datatype'])]
+                    if fbasetype == 'A':
+                        log.debug("String to integer conversion required for %s -> %s.", fcol, col['column_name'])
+                        width = int(str(old[fcol].dtype).split(old[fcol].dtype.kind)[1])
+                        blank = ' '*width
+                        w = np.nonzero(old[fcol] == blank)[0]
+                        if len(w) > 0:
+                            log.debug("old['%s'][old['%s'] == blank] = blank[0:%d] + '0'",
+                                      fcol, fcol, width - 1)
+                            old[fcol][w] = blank[0:(width-1)] + '0'
+                        log.debug("test_old = old['%s'].astype(np.int64)", fcol)
+                        try:
+                            test_old = old[fcol].astype(np.int64)
+                        except OverflowError:
+                            log.debug("Attempting string to quasi-unsigned integer conversion for %s -> %s.",
+                                      fcol, col['column_name'])
+                            uold = old[fcol].astype(np.uint64)
+                            hi = np.nonzero(uold >= 2**63)[0]
+                            lo = np.nonzero(uold < 2**63)[0]
+                            test_old = np.zeros(uold.shape, dtype=np.int64)
+                            test_old[lo] = uold[lo]
+                            test_old[hi] = (uold[hi] - 2**63).astype(np.int64) - 2**63
+                    else:
+                        if index is not None:
+                            test_old = old[fcol][:, index]
+                        else:
+                            test_old = old[fcol]
+                    if ((test_old >= -limit) & (test_old <= limit - 1)).all():
+                        if (fbasetype, col['datatype']) == ('A', 'bigint'):
+                            log.debug("new['%s'] = test_old  # quasi-unsigned integer", col['column_name'])
+                            new[col['column_name']] = test_old
+                        else:
+                            if index is not None:
+                                log.debug("new['%s'] = old['%s'][:, %d].astype(%s)", col['column_name'], fcol, index, str(np_map[col['datatype']]))
+                                new[col['column_name']] = old[fcol][:, index].astype(np_map[col['datatype']])
+                            else:
+                                log.debug("new['%s'] = old['%s'].astype(%s)", col['column_name'], fcol, str(np_map[col['datatype']]))
+                                new[col['column_name']] = old[fcol].astype(np_map[col['datatype']])
+                    else:
+                        msg = "Values too large for safe data type conversion for %s (%s) -> %s (%s)!"
+                        log.error(msg, fcol, fbasetype, col['column_name'], col['datatype'])
+                        raise ValueError(msg % (fcol, fbasetype, col['column_name'], col['datatype']))
+                else:
+                    msg = "No safe data type conversion possible for %s (%s) -> %s (%s)!"
+                    log.error(msg, fcol, fbasetype, col['column_name'], col['datatype'])
+                    raise ValueError(msg % (fcol, fbasetype, col['column_name'], col['datatype']))
+            if fbasetype in ('D', 'E'):
+                new[col['column_name']][~np.isfinite(new[col['column_name']])] = -9999.0
+        log.debug("new.write('%s')", out)
+        new.write(out)
+        return out
 
 
 def get_options():
@@ -249,13 +520,19 @@ def get_options():
                         metavar='TEXT',
                         default='Sloan Digital Sky Survey Data Relase 14',
                         help='Short description of the schema.')
+    parser.add_argument('-E', '--no-ecliptic', dest='ecliptic', action='store_false',
+                        help='Do not add ecliptic coordinates.')
     parser.add_argument('-e', '--extension', dest='hdu', metavar='N',
                         type=int, default=1,
                         help='Read data from FITS HDU N (default %(default)s).')
+    parser.add_argument('-G', '--no-galactic', dest='galactic', action='store_false',
+                        help='Do not add galactic coordinates.')
     parser.add_argument('-j', '--output-json', dest='output_json', metavar='FILE',
                         help='Write table metadata to FILE.')
     parser.add_argument('-k', '--keep', action='store_true',
                         help='Do not overwrite any existing intermediate files.')
+    parser.add_argument('-l', '--log', dest='log', metavar='FILE',
+                        help='Log operations to FILE.')
     parser.add_argument('-m', '--merge', dest='merge_json', metavar='FILE',
                         help='Merge metadata in FILE into final metadata output.')
     parser.add_argument('-o', '--output-sql', dest='output_sql', metavar='FILE',
@@ -290,6 +567,8 @@ def main():
                                           "%s.%s.sql" % (options.schema, options.table))
     if options.output_json is None:
         options.output_json = options.output_sql.replace('sql', 'json')
+    if options.log is None:
+        options.log = options.output_sql.replace('sql', 'log')
     try:
         sdss = SDSS(options.schema, options.table,
                     description=options.description,
@@ -300,7 +579,7 @@ def main():
         #
         print(str(e))
         return 1
-    sdss.configureLog(options.verbose)
+    sdss.configureLog(options.log, options.verbose)
     log = sdss.logName('sdss.main')
     # ts = datetime.utcnow().replace(tzinfo=utc).strftime('%Y-%m-%dT%H:%M:%S %Z')
     log.debug("options.fits = '%s'", options.fits)
@@ -309,12 +588,16 @@ def main():
     log.debug("options.table = '%s'", options.table)
     log.debug("options.output_sql = '%s'", options.output_sql)
     log.debug("options.output_json = '%s'", options.output_json)
+    log.debug("options.log = '%s'", options.log)
     #
     # Preprocess the FITS file.
     #
+    sdss.customSTILTS(options.config)
     try:
         dlfits = sdss.addDLColumns(options.fits, ra=options.ra,
-                                   overwrite=(not options.keep))
+                                   overwrite=(not options.keep),
+                                   ecliptic=options.ecliptic,
+                                   galactic=options.galactic)
     except ValueError as e:
         log.error(str(e))
         return 1
@@ -326,9 +609,11 @@ def main():
     #
     # Map the FITS columns to table columns.
     #
+    sdss.fixNOFITS(options.config)
+    sdss.fixMapping(options.config)
     try:
         sdss.mapColumns()
-    except KeyError as e:
+    except KeyError as k:
         return 1
     #
     # Fix any table definition problems and sort the columns.
@@ -356,4 +641,7 @@ def main():
                                   overwrite=(not options.keep))
     except ValueError as e:
         return 1
+    # except Exception as e:
+    #     log.error(str(e))
+    #     return 2
     return 0