+++ digestor/test/test_sdss.py
@@ -6,8 +6,11 @@ import unittest
 import unittest.mock as mock
 from tempfile import NamedTemporaryFile
 
+import numpy as np
+
 from ..sdss import SDSS, get_options
 from .utils import DigestorCase
+
 
 class TestSDSS(DigestorCase):
     """Test digestor.sdss.
@@ -109,6 +112,43 @@ class TestSDSS(DigestorCase):
         self.assertEqual(d['unit'], 'nanomaggies')
         self.assertEqual(d['description'], 'Random column.')
 
+    def test_fix_nofits(self):
+        """Test adjustment of missing columns with YAML configuration.
+        """
+        yaml = """sdss:
+    spectra:
+        NOFITS:
+            glon: drop
+            glat: drop
+        """
+        self.assertEqual(len(self.sdss.NOFITS), 0)
+        with NamedTemporaryFile('w+') as f:
+            f.write(yaml)
+            f.seek(0)
+            self.sdss.fixNOFITS(f.name)
+        self.assertEqual(self.sdss.NOFITS['glon'], 'drop')
+        self.assertEqual(self.sdss.NOFITS['glat'], 'drop')
+        self.sdss.table = 'foo'
+        self.sdss.fixNOFITS(f.name)
+
+    def test_fix_mapping(self):
+        """Test adjustment of column mapping with YAML configuration.
+        """
+        yaml = """sdss:
+    spectra:
+        mapping:
+            glon: L
+            glat: B
+        """
+        with NamedTemporaryFile('w+') as f:
+            f.write(yaml)
+            f.seek(0)
+            self.sdss.fixMapping(f.name)
+        self.assertEqual(self.sdss.mapping['glon'], 'L')
+        self.assertEqual(self.sdss.mapping['glat'], 'B')
+        self.sdss.table = 'foo'
+        self.sdss.fixMapping(f.name)
+
     def test_map_columns(self):
         """Test mapping of FITS columns to SQL columns.
         """
@@ -135,6 +175,18 @@ class TestSDSS(DigestorCase):
                                             "description": "g ivar",
                                             "unit": "", "ucd": "", "utype": "",
                                             "datatype": "real", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "no_fits_keep",
+                                            "description": "keeper",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "real", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "no_fits_drop",
+                                            "description": "dropper",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "real", "size": 1,
                                             "principal": 0, "indexed": 0, "std": 0}]
         self.sdss.mapping = {'mag_u': 'MAG[0]', 'mag_g': 'MAG[1]',
                              'magivar_u': 'MAGIVAR[0]', 'magivar_g': 'MAGIVAR[1]'}
@@ -144,6 +196,7 @@ class TestSDSS(DigestorCase):
                           'nest4096': 'J', 'MAG': '2E',
                           'MAG_IVAR': '2E',
                           'FOOBAR': '16A'}
+        self.sdss.NOFITS = {'no_fits_keep': 'defer', 'no_fits_drop': 'drop'}
         self.sdss.mapColumns()
         final_mapping = {'mag_u': 'MAG[0]', 'mag_g': 'MAG[1]',
                          'magivar_u': 'MAG_IVAR[0]', 'magivar_g': 'MAG_IVAR[1]',
@@ -163,7 +216,7 @@ class TestSDSS(DigestorCase):
                                             "description": "g flux",
                                             "unit": "", "ucd": "", "utype": "",
                                             "datatype": "real", "size": 1,
-                                            "principal": 0, "indexed": 0, "std": 0},]
+                                            "principal": 0, "indexed": 0, "std": 0}]
         self.sdss.mapping = {'mag_u': 'MAG[0]', 'mag_g': 'MAG[1]',
                              'magivar_u': 'MAGIVAR[0]', 'magivar_g': 'MAGIVAR[1]',
                              'flux_u': 'FLUX[0]', 'flux_g': 'FLUX[1]'}
@@ -176,13 +229,227 @@ class TestSDSS(DigestorCase):
                                             "description": "z",
                                             "unit": "", "ucd": "", "utype": "",
                                             "datatype": "real", "size": 1,
-                                            "principal": 0, "indexed": 0, "std": 0},]
+                                            "principal": 0, "indexed": 0, "std": 0}]
         self.sdss.mapping = {'mag_u': 'MAG[0]', 'mag_g': 'MAG[1]',
                              'magivar_u': 'MAGIVAR[0]', 'magivar_g': 'MAGIVAR[1]',
                              'flux_u': 'FLUX[0]', 'flux_g': 'FLUX[1]'}
         with self.assertRaises(KeyError) as e:
             self.sdss.mapColumns()
         self.assertEqual(e.exception.args[0], 'Could not find a FITS column corresponding to z!')
+        self.sdss.tapSchema['columns'] += [{"table_name": self.table,
+                                            "column_name": "no_fits_error",
+                                            "description": "error",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "real", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0}]
+        self.sdss.FITS['Z'] = 'E'
+        self.sdss.NOFITS = {'no_fits_keep': 'defer', 'no_fits_drop': 'drop', 'no_fits_error': 'foobar'}
+        with self.assertRaises(KeyError) as e:
+            self.sdss.mapColumns()
+        self.assertEqual(e.exception.args[0], 'Unknown NOFITS instruction: no_fits_error!')
+
+    def test_process_fits(self):
+        """Test processing of SDSS-specific FITS file for loading.
+        """
+        self.sdss.tapSchema['columns'] += [{"table_name": self.table,
+                                            "column_name": "mag_u",
+                                            "description": "u Magnitude",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "real", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "mag_g",
+                                            "description": "g Magnitude",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "real", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "magivar_u",
+                                            "description": "u ivar",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "double", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "magivar_g",
+                                            "description": "g ivar",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "double", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "objid",
+                                            "description": "id",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "bigint", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "bigobjid",
+                                            "description": "id",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "bigint", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "smallid",
+                                            "description": "id",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "smallint", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "unsafe",
+                                            "description": "unsafe",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "integer", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "unsafe2",
+                                            "description": "unsafe",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "smallint", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "small_bit",
+                                            "description": "unsafe",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "smallint", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "small_flags_u",
+                                            "description": "unsafe",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "smallint", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "flags",
+                                            "description": "unsafe",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "bigint", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "flags_u",
+                                            "description": "unsafe",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "bigint", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "no_fits_keep",
+                                            "description": "unsafe",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "integer", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0},
+                                           {"table_name": self.table,
+                                            "column_name": "no_fits_drop",
+                                            "description": "unsafe",
+                                            "unit": "", "ucd": "", "utype": "",
+                                            "datatype": "integer", "size": 1,
+                                            "principal": 0, "indexed": 0, "std": 0}]
+        i = self.sdss.columnIndex('nest4096')
+        u = self.sdss.columnIndex('unsafe')
+        self.sdss.tapSchema['columns'][i]['datatype'] = 'smallint'
+        self.sdss.FITS = {'elon': 'D', 'elat': 'D',
+                          'glon': 'E', 'glat': 'E',
+                          'htm9': 'J', 'ring256': 'J',
+                          'nest4096': 'J',
+                          'random_id': 'E',
+                          'mag': '2E', 'magivar': '2E',
+                          'objid': '16A',
+                          'bigobjid': '20A',
+                          'smallid': '3A',
+                          'foobar': '16A',
+                          'small_bit': 'J',
+                          'small_bits': '5J',
+                          'objc_flags': 'J',
+                          'objc_flags2': 'J',
+                          'flags': '5J',
+                          'flags2': '5J',
+                          'unsafe': 'K',
+                          'unsafe2': 'J'}
+        self.sdss.NOFITS = {'no_fits_keep': 'defer', 'no_fits_drop': 'drop'}
+        for k in self.sdss.FITS:
+            self.sdss.mapping[k] = k
+        self.sdss.mapping['mag_u'] = 'mag[0]'
+        self.sdss.mapping['mag_g'] = 'mag[1]'
+        self.sdss.mapping['magivar_u'] = 'magivar[0]'
+        self.sdss.mapping['magivar_g'] = 'magivar[1]'
+        self.sdss.mapping['small_flags_u'] = 'small_bits[0]'
+        self.sdss.mapping['flags'] = 'objc_flags'
+        self.sdss.mapping['flags_u'] = 'flags[0]'
+        self.sdss._inputFITS = 'foo.fits'
+        dummy_values = {'elon': np.ones((5,), dtype=np.float64),
+                        'elat': np.ones((5,), dtype=np.float64),
+                        'glon': np.ones((5,), dtype=np.float32),
+                        'glat': np.ones((5,), dtype=np.float32),
+                        'htm9': np.ones((5,), dtype=np.int32),
+                        'ring256': np.ones((5,), dtype=np.int32),
+                        'nest4096': np.ones((5,), dtype=np.int32),
+                        'random_id': np.ones((5,), dtype=np.float32),
+                        'mag': np.ones((5, 2), dtype=np.float32),
+                        'magivar': np.ones((5, 2), dtype=np.float32),
+                        'objid': np.array([' '*15 + '1']*4 + [' '*16], dtype='U16'),
+                        'bigobjid': np.array(['9223372036854775808']*3 + ['18446744073709551615']*2, dtype='U20'),
+                        'smallid': np.array(['123']*3 + ['   ']*2, dtype='U3'),
+                        'foobar': np.array([' '*16]*5, dtype='U16'),
+                        'small_bit': np.ones((5,), dtype=np.int32),
+                        'small_bits': np.ones((5, 5), dtype=np.int32),
+                        'objc_flags': np.ones((5,), dtype=np.int32),
+                        'objc_flags2': np.ones((5,), dtype=np.int32),
+                        'flags': np.ones((5, 5), dtype=np.int32),
+                        'flags2': np.ones((5, 5), dtype=np.int32),
+                        'unsafe': np.ones((5,), dtype=np.int64),
+                        'unsafe2': np.ones((5,), dtype=np.int32) + 2**15}
+        #
+        # Raise an unsafe error.
+        #
+        with mock.patch('digestor.sdss.Table') as T:
+            t = T.read.return_value = mock.MagicMock()
+            t.__getitem__.side_effect = lambda key: dummy_values[key.lower()]
+            t.colnames = [k.upper() for k in dummy_values.keys()]
+            with self.assertRaises(ValueError) as e:
+                self.sdss.processFITS()
+            self.assertEqual(e.exception.args[0], 'No safe data type conversion possible for unsafe (K) -> unsafe (integer)!')
+        del dummy_values['unsafe']
+        del self.sdss.FITS['unsafe']
+        del self.sdss.mapping['unsafe']
+        del self.sdss.tapSchema['columns'][u]
+        u2 = self.sdss.columnIndex('unsafe2')
+        #
+        # Try again.
+        #
+        with mock.patch('digestor.sdss.Table') as T:
+            t = T.read.return_value = mock.MagicMock()
+            t.__getitem__.side_effect = lambda key: dummy_values[key.lower()]
+            t.colnames = [k.upper() for k in dummy_values.keys()]
+            with self.assertRaises(ValueError) as e:
+                self.sdss.processFITS()
+            self.assertEqual(e.exception.args[0], 'Values too large for safe data type conversion for unsafe2 (J) -> unsafe2 (smallint)!')
+        del dummy_values['unsafe2']
+        del self.sdss.FITS['unsafe2']
+        del self.sdss.mapping['unsafe2']
+        del self.sdss.tapSchema['columns'][u2]
+        #
+        # Try again.
+        #
+        with mock.patch('digestor.sdss.Table') as T:
+            t = T.read.return_value = mock.MagicMock()
+            t.__getitem__.side_effect = lambda key: dummy_values[key.lower()]
+            t.colnames = [k.upper() for k in dummy_values.keys()]
+            out = self.sdss.processFITS()
+        self.assertEqual(out, '{0.schema}.{0.table}.fits'.format(self))
+        #
+        # Check overwrite
+        #
+        with mock.patch('os.path.exists') as ex:
+            ex.return_value = True
+            out = self.sdss.processFITS()
+            ex.assert_called_with(out)
+        with mock.patch('os.path.exists') as ex:
+            with mock.patch('os.remove') as rm:
+                with mock.patch('digestor.sdss.Table') as T:
+                    t = T.read.return_value = mock.MagicMock()
+                    t.__getitem__.side_effect = lambda key: dummy_values[key.lower()]
+                    t.colnames = [k.upper() for k in dummy_values.keys()]
+                    ex.return_value = True
+                    out = self.sdss.processFITS(overwrite=True)
+            rm.assert_called_with(out)
+            ex.assert_called_with(out)
 
 
 def test_suite():